move join table inference
